django-starter/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==4.2.11
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.3
sqlparse==0.4.4
typing_extensions==4.10.0
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.2